debian: Use debconf to ask for privacy permissions
This is the only possible way because Debian allows for noninteractive
installers, such as Ubuntu Software Center.  With such installers we
cannot ask questions on the standard input because the standard input
is a pipe.  And we cannot open a browser window because there is no
display available in such context.

See also <https://github.com/bassosimone/neubot-privacy-deb>.