README.rst: rewrite
This patch basically rewrites README.rst for the better and is joint
work of Felipe Busnello and me.  Credit is here given to Felipe because
he did about 66% of the work.  More in detail, Felipe envisioned and
assembled the initial text, which then was a bit rehashed by me.
In particular, I splitted again the text in many paragrahs, which
was Felipe's initial take.  Since just one paragraph makes the text
too dense.

I don't see a need to provide here all the patches and I am instead
collapsing them into a single commit.

--
Simone Basso